backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
celery>=5.4.0
redis>=5.0.0
playwright>=1.48.0
anthropic>=0.40.0
supabase>=2.10.0
tavily-python>=0.5.0
weasyprint>=62.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic-settings>=2.5.0
jinja2>=3.1.0
loguru>=0.7.0
sentry-sdk[fastapi]>=2.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
